+++ CHANGELOG.md
@@ -14,3 +14,4 @@
 - Load JSON file from data service using HttpClient
 - Show superheroes data in an Angular Material table
 - Sort values by each column
+- Move table to a new component into elements folder